middleware-v1/requirements.txt
# Follow alphabetical order
aws-lambda-powertools
boto3~=1.26.33
botocore~=1.29.33
fastapi~=0.88.0
google-search-results
marshmallow_dataclass
path~=16.6.0
uvicorn
mangum~=0.17.0
httpx~=0.23.1
validators~=0.20.0
pip~=22.0.2
attrs~=22.2.0
wheel~=0.37.1
pytest~=7.2.0
pluggy~=1.0.0
exceptiongroup~=1.1.0
iniconfig~=1.1.1
setuptools~=59.6.0
sentry_sdk
marshmallow~=3.19.0
requests~=2.28.1
Jinja2~=3.1.2
fuzzywuzzy~=0.18.0
inflect
tldextract
cachetools
psycopg2-binary
aws-psycopg2
aws_xray_sdk
